+++ src/main/java/ba/unsa/etf/si/payment/controller/MoneyTransferController.java
@@ -4,10 +4,10 @@ import ba.unsa.etf.si.payment.model.ApplicationUser;
 import ba.unsa.etf.si.payment.model.BankAccount;
 import ba.unsa.etf.si.payment.model.BankAccountUser;
 import ba.unsa.etf.si.payment.model.MoneyTransfer;
-import ba.unsa.etf.si.payment.request.Transfer.MoneyTransferInRequest;
-import ba.unsa.etf.si.payment.request.Transfer.MoneyTransferRequest;
-import ba.unsa.etf.si.payment.response.Transfer.MoneyTransferResponse;
-import ba.unsa.etf.si.payment.response.Transfer.TransferResponse;
+import ba.unsa.etf.si.payment.request.transferRequest.MoneyTransferInRequest;
+import ba.unsa.etf.si.payment.request.transferRequest.MoneyTransferRequest;
+import ba.unsa.etf.si.payment.response.transferResponse.MoneyTransferResponse;
+import ba.unsa.etf.si.payment.response.transferResponse.TransferResponse;
 import ba.unsa.etf.si.payment.security.CurrentUser;
 import ba.unsa.etf.si.payment.security.UserPrincipal;
 import ba.unsa.etf.si.payment.service.ApplicationUserService;
