+++ src/main/java/ba/unsa/etf/si/payment/request/qrCodes/StaticQRRequest.java
@@ -1,4 +1,4 @@
-package ba.unsa.etf.si.payment.request.QRCodes;
+package ba.unsa.etf.si.payment.request.qrCodes;
 
 import javax.validation.constraints.NotBlank;
