+++ src/main/java/ba/unsa/etf/si/payment/request/transferRequest/MoneyTransferRequest.java
@@ -1,4 +1,4 @@
-package ba.unsa.etf.si.payment.request.Transfer;
+package ba.unsa.etf.si.payment.request.transferRequest;
 
 import javax.validation.constraints.NotBlank;
